Create 4.7 Deriving New Variables
These queries create variables and groupings for Instarcart data.

If-Then statement and If-Then Statement loc() used to sort products by low-high price range. For Loop generated counts to observe orders by day of the week (dow). If-Then statement used to generate the busiest day of the week for Instacart.
Further queries used to count orders by the hour of the day and an If-Then statement used to around the busiest hour of the day by order frequency.